requirements.txt
Django>=4.2,<6
djangorestframework>=3.15,<4
psycopg[binary]>=3.1,<4
gunicorn>=21,<22
